:alembic: ci: Initial commit, add .travis.yml (#3)
This commit sets up a basic Continuous Integration test suite. It runs a
YAML linter and an Ansible Playbook linter to catch common mistakes. The
results of the CI tests are reported back in a Pull Request or the
branch status.

This will make it easier to validate and test new changes for common
mistakes and errors.

Signed-off-by: Justin W. Flory <git@jwf.io>