Provision Anvil playbook for a Spigot server role
This commit updates the server inventory from a TOML format to a YAML
file. We will use advanced host metadata, so it is more convenient to
store it in this format. As proof-of-concept, the server alias is now
`anvil`, where the `ansible_host` is defined as `anvil.ccmc.pw` and a
group of host admins are specified.

Signed-off-by: Justin W. Flory <git@jwf.io>