:package: spigot: Lock MC version in BuildTools, use var for file owner
This commit cleans up how the binaries get built. Now, BuildTools is
fixed to a specific version identifier, so it will always build Spigot
for a specific version. This was intentional, because `1.16.2` was not
offered on the `latest` release stream at time of commit.

Using a variable for the owner/group is better practice anyways.

Signed-off-by: Justin W. Flory <git@jwf.io>