:wrench: :broom: spigot: Convert spigot-server directory to variable
This commit changes `/var/games/spigot-server/` directory into a
variable. This makes it easier to manage the path for the Spigot server,
and will also be used in a subsequent change that fixes the issue with
the server running in the root directory.

But this is just the prep work for the real fix!

Signed-off-by: Justin W. Flory <git@jwf.io>