:beetle: :scream: spigot: Set WorkingDirectory to avoid perms problems (#8)
This commit adds a `WorkingDirectory` line into the systemd service file
used for starting, stopping, and restarting the server. This was a
mistake I made while setting up the systemd service.

Originally, the systemd service had the full path to the `spigot.jar` in
the `ExecStart` statement. But that caused the process to be started
from the root directory, meaning all of the server files and data were
being written to the root directory. Oh no!

So, this sets `WorkingDirectory` to the path of the Spigot server, and
it makes the path to `spigot.jar` local. So this way, the server should
never start if it is not in the right directory.

I tested this change on `anvil.ccmc.pw` and I confirmed that it worked.
It also picked up the plugins I had uploaded previously. So, I think
this will get everything working and in good shape! Sorry about the
bumpy road in the beginning.

Signed-off-by: Justin W. Flory <git@jwf.io>