:beetle: :wrench: nginx: Fix broken cert path for TLS certificates (#9)
This commit fixes a bug in the nginx config by changing the Ansible
variable used for the server hostname.

Previously, `inventory_hostname` was used, but that evaluates to `anvil`
per the `inventory/inventory` file. This commit changes
`inventory_hostname` to `ansible_host`, which evaluates to the fully-
qualified domain name (FQDN) for the server: `anvil.ccmc.pw`.

This is the path of the LetsEncrypt/Certbot certificates. It is not
clear to me why this changed or how it broke suddenly, but this change
fixes the issue. At commit time, I ran this change into production and
verified it solved the issue.

cc: @JMW18

Signed-off-by: Justin W. Flory <git@jwf.io>