:new: system/user-management: Initial commit
This commit adds a new role to manage users and privilege-escalation on
the Spigot game host. Most notably, it creates a custom sudoers policy
for the `spigot-admin` group and delegates specific privileges to all
members of the group.

This is a security hardening technique and clearly defines a policy for
Spigot game server admins. This also enables the possibility of adding
other sysadmins to the team at fairly low risk.

cc: @JMW18

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>

Fix tag syntax in meta file

Signed-off-by: Justin W. Flory (he/him) <git@jwf.io>